Presence live updates (transient data) *Don't do such big commit again*

+++ wave/src/main/java/org/swellrt/beta/client/platform/web/editor/CaretInfo.java
@@ -13,7 +13,7 @@ public interface CaretInfo {
   String getParticipant();
 
   @JsProperty
-  long getTimestamp();
+  int getTimestamp();
 
   @JsProperty
   int getPosition();
